fixed share/delete bubble down calls
